+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changes
 
+- read datenkompass default values from mapping
+
 ### Deprecated
 
 ### Removed
@@ -21,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - empty list handling in datenkompass
+- odk variable identifier in primary source
 
 ### Security